+++ changelog.md
@@ -1,7 +1,9 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-# [0.0.71] - In progress
+# [0.0.72] - In progress
+
+# [0.0.71] - 2024-05-20
 * add necessary metadata to files Builder3
 * add buildBookTemplate for Builder3
 * add temp folder concept to Builder3